+++ src/kOS.Safe.Test/Structure/NoArgsSuffixTest.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using NUnit.Framework;
 using kOS.Safe.Encapsulation.Suffixes;
+using NUnit.Framework;
 
-namespace kOS.Safe.Test
+namespace kOS.Safe.Test.Structure
 {
     [TestFixture]
     public class NoArgsSuffixTest
