+++ src/kOS.Safe.Test/kOS.Safe.Test.csproj
@@ -45,19 +45,19 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClampSetSuffixTest.cs" />
-    <Compile Include="LexiconTest.cs" />
-    <Compile Include="ListValueTest.cs" />
-    <Compile Include="NoArgsSuffixTest.cs" />
+    <Compile Include="Structure\ClampSetSuffixTest.cs" />
+    <Compile Include="Collections\LexiconTest.cs" />
+    <Compile Include="Collections\ListValueTest.cs" />
+    <Compile Include="Structure\NoArgsSuffixTest.cs" />
     <Compile Include="Opcode\FakeCpu.cs" />
     <Compile Include="Opcode\OpcodeGetIndexText.cs" />
     <Compile Include="Opcode\OpcodeSetIndexTest.cs" />
-    <Compile Include="SetSuffixTest.cs" />
-    <Compile Include="StructureSuffixIntegrationTest.cs" />
-    <Compile Include="StructureTest.cs" />
+    <Compile Include="Structure\SetSuffixTest.cs" />
+    <Compile Include="Structure\StructureSuffixIntegrationTest.cs" />
+    <Compile Include="Structure\StructureTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestLogger.cs" />
-    <Compile Include="TwoArgsVoidSuffixTest.cs" />
+    <Compile Include="Structure\TwoArgsVoidSuffixTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\kOS.Safe\kOS.Safe.csproj">
